Make all sounds work, not just manim-voiceover
