+++ manim_slides/slide/__init__.py
@@ -9,7 +9,6 @@ __all__ = [
 
 import os
 import sys
-import importlib.util
 
 
 class ManimApiNotFoundError(ImportError):
@@ -49,14 +48,8 @@ if not os.environ.get(FORCE_MANIM_API):
 
 MANIM: bool = API_NAME == "manim"
 MANIMGL: bool = API_NAME == "manimlib"
-MANIM_VOICEOVER: bool = importlib.util.find_spec("manim_voiceover") is not None
 
-if MANIM and MANIM_VOICEOVER:
-    try:
-        from .manimvoiceover import Slide, ThreeDSlide
-    except ImportError as e:
-        raise ManimApiNotFoundError from e
-elif MANIM:
+if MANIM:
     try:
         from .manim import Slide, ThreeDSlide
     except ImportError as e:
