+++ pyproject.toml
@@ -64,7 +64,6 @@ pyqt6 = ["pyqt6>=6.7.0"]
 pyqt6-full = ["manim-slides[full,pyqt6]"]
 pyside6 = ["pyside6>=6.6.1"]
 pyside6-full = ["manim-slides[full,pyside6]"]
-manim-voiceover = ["manim-voiceover[gtts,recorder]>=0.3.7"]
 sphinx-directive = ["docutils>=0.20.1", "manim-slides[manim]"]
 tests = [
   "manim-slides[full,manimgl,pyqt6,pyside6,sphinx-directive]",
